requirements.txt
aiofiles==0.6.0
autopep8==1.5.7
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
ecdsa==0.14.1
fastapi==0.63.0
fastapi-login==1.5.3
firebase-admin==4.5.3
gcloud==0.18.3
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
h11==0.12.0
httplib2==0.19.0
idna==2.10
Jinja2==2.11.3
jwcrypto==0.8
MarkupSafe==1.1.1
msgpack==1.0.2
ndg-httpsclient==0.5.1
numpy==1.20.1
oauth2client==4.1.3
packaging==20.9
passlib==1.7.4
proto-plus==1.18.1
protobuf==3.15.6
pusher==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.1
PyJWT==2.0.1
pymongo==3.11.3
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-jose==3.2.0
python-jwt==3.3.0
python-multipart==0.0.5
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
